Rename files, update config
