+++ babel.config.js
@@ -1,4 +1,11 @@
 module.exports = {
-  presets: ['@babel/preset-env'],
+  presets: [
+    [
+      '@babel/preset-env',
+      {
+        targets: '> 0.25%, not dead',
+      },
+    ],
+  ],
   plugins: ['@babel/plugin-proposal-class-properties'],
 };
