+++ package.json
@@ -51,7 +51,6 @@
     "optimize-css-assets-webpack-plugin": "^5.0.3",
     "prettier": "^1.18.2",
     "style-loader": "^1.0.0",
-    "terser-webpack-plugin": "^2.1.2",
     "watchify": "^3.11.1",
     "webpack": "^4.41.0",
     "webpack-cli": "^3.3.9"
