+++ spec/spec/module.spec.js
